transactions formatted to look better

+++ db/spending_tracker.sql
@@ -14,7 +14,7 @@ CREATE TABLE merchants (
 
 CREATE TABLE transactions (
     id SERIAL PRIMARY KEY,
-    amount FLOAT,
+    amount DECIMAL,
     merchant_id INT REFERENCES merchants(id),
     tag_id INT REFERENCES tags(id),
     created_at timestamp default current_timestamp
